requirements.txt
dislash.py
discord.py